split the cli handling and logic for serve command
the serve logic and the registry helpers have been moved to the serve module
